feat(frontend): mysql webconsole支持角色选择 #12657 # Reviewed, transaction id: 56836
