+++ dbm-ui/frontend/src/views/db-manage/common/webconsole/components/console-panel/tendbcluster/Index.vue
@@ -21,8 +21,7 @@
   import type { queryAllTypeCluster } from '@services/source/dbbase';
 
   import ConsoleInput from '../components/ConsoleInput.vue';
-
-  import RenderMessage from './components/RenderMessage.vue';
+  import RenderMessage from '../mysql/components/RenderMessage.vue';
 
   interface Props {
     charset: string;
@@ -38,9 +37,8 @@
   const consoleInputRef = ref<typeof ConsoleInput>();
 
   const roleDisplay = {
-    backend_master: 'Master',
-    backend_slave: 'Slave',
-    orphan: 'Orphan',
+    spider_master: 'Spider Master',
+    spider_slave: 'Spider Slave',
   };
 
   const placeholder = computed(() => `${props.cluster.immute_domain}[${roleDisplay[props.role]}] > `);
